Added Unit Tests

+++ BookStore/BookStore.Services/ShoppingCartService.cs
@@ -13,20 +13,18 @@ namespace BookStore.Services
     public class ShoppingCartService : IShoppingCartService
     {
         private readonly BookStoreDbContext db;
-        private readonly IBookService bookService;
         private readonly IUserService userService;
 
-        public ShoppingCartService(BookStoreDbContext db ,IBookService bookService,IUserService userService)
+        public ShoppingCartService(BookStoreDbContext db, IUserService userService)
         {
             this.db = db;
-            this.bookService = bookService;
             this.userService = userService;
         }
         public async Task<bool> AddBookToShoppingCartAsync(int id, string username)
         {
             var book = this.db.Books.Find(id);
             var user = await this.userService.GetByUsernameAsync(username);
-            var shoppingCart = this.db.ShoppingCarts.Find(user.ShoppingCartId);
+            var shoppingCart = this.db.ShoppingCarts.Find(user?.ShoppingCartId);
 
             if (book == null || user == null)
             {
@@ -137,25 +135,25 @@ namespace BookStore.Services
             return result > 0;
         }
 
-        private ShoppingCartBook GetShoppingCartBook(int bookId, int shoppingCartId)
-        {
-            return this.db.ShoppingCartBooks.FirstOrDefault(shb => shb.ShoppingCartId == shoppingCartId && shb.BookId == bookId);
-        }
-
         public async Task<bool> RemoveBooksFromShoppingCartAsync(string username)
         {
             var user = await this.userService.GetByUsernameAsync(username);
 
-            //if (user == null)
-            //{
-            //    return;
-            //}
+            if (user == null)
+            {
+                return false;
+            }
 
             var shoppingCartBooks = this.db.ShoppingCartBooks.Where(shb => shb.ShoppingCartId == user.ShoppingCartId);
 
             this.db.ShoppingCartBooks.RemoveRange(shoppingCartBooks);
             int result = this.db.SaveChanges();
             return result > 0;
+        }
+
+        private ShoppingCartBook GetShoppingCartBook(int bookId, int shoppingCartId)
+        {
+            return this.db.ShoppingCartBooks.FirstOrDefault(shb => shb.ShoppingCartId == shoppingCartId && shb.BookId == bookId);
         }
     }
 }
